moved classic and whosthatpokemon routes into nested folder for shared layout
