+++ src/app/(pokemon)/(modes)/classic/_components/DailyGame.tsx
