+++ src/app/(pokemon)/(modes)/classic/_components/Gamebox.tsx
