+++ src/app/(pokemon)/(modes)/classic/page.tsx
