+++ src/app/(pokemon)/(modes)/whosthatpokemon/_components/Gamebox.tsx
