+++ src/app/(pokemon)/(modes)/whosthatpokemon/_components/ImagePanel.tsx
