+++ src/app/(pokemon)/_components/GameWrapper.tsx
@@ -14,8 +14,8 @@ import { useAtomValue, useAtom, useSetAtom } from "jotai";
 import { useHydrateAtoms } from "jotai/utils";
 import { usePathname, useSearchParams } from "next/navigation";
 import { useEffect } from "react";
-import ClassicGamebox from "../classic/_components/Gamebox";
-import WhosThatPokemonGamebox from "../whosthatpokemon/_components/Gamebox";
+import ClassicGamebox from "../(modes)/classic/_components/Gamebox";
+import WhosThatPokemonGamebox from "../(modes)/whosthatpokemon/_components/Gamebox";
 import PokemonTypes from "./PokemonTypes";
 import PokemonSearch from "./PokemonSearch";
 import { getLocalTimeZone, today } from "@internationalized/date";
